登录校验

+++ market-wx-service/src/main/java/com/lihebin/market/wx/domain/AdminReq.java
@@ -49,6 +49,6 @@ public class AdminReq {
     /**
      * 角色列表
      */
-    private List<Integer> roleIds;
+    private List<Long> roleIds;
 
 }
